Move `RangeOnPath` to `path`
